Moved getElement() implementation to GetElementBehavior classes
