--- src/main/java/com/softwareonpurpose/uinavigator/behavior/getelement/GetElementDirectly.java
@@ -1,14 +1,0 @@
-package com.softwareonpurpose.uinavigator.behavior.getelement;
-
-import org.openqa.selenium.WebElement;
-
-public class GetElementDirectly implements GetWebElementBehavior {
-    public static GetElementDirectly getInstance() {
-        return new GetElementDirectly();
-    }
-
-    @Override
-    public WebElement execute() {
-        return null;
-    }
-}
