--- src/main/java/com/softwareonpurpose/uinavigator/behavior/getelement/GetElementFromList.java
@@ -1,14 +1,0 @@
-package com.softwareonpurpose.uinavigator.behavior.getelement;
-
-import org.openqa.selenium.WebElement;
-
-public class GetElementFromList implements GetWebElementBehavior {
-    public static GetElementFromList getInstance() {
-        return new GetElementFromList();
-    }
-
-    @Override
-    public WebElement execute() {
-        return null;
-    }
-}
